Added multiples arbitrary; moved some arbitraries to proper package
